+++ base/src/ipfixcol.c
@@ -40,10 +40,12 @@
 #if HAVE_CONFIG_H
 #include <pkgconfig.h>
 #endif
+
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
 #include <errno.h>
+#include <limits.h>
 #include <stdbool.h>
 #include <string.h>
 #include <sys/types.h>
@@ -57,10 +59,10 @@
 #include <sys/prctl.h>
 
 #include <libxml/parser.h>
+#include <libxml/tree.h>
 #include <libxml/xpath.h>
 
 #include <ipfixcol.h>
-#include <libxml/tree.h>
 #include "intermediate_process.h"
 #include "config.h"
 #include "preprocessor.h"
@@ -75,7 +77,6 @@
  *
  */
 
-
 /** Acceptable command-line parameters */
 #define OPTSTRING "c:dhv:Vsr:i:S:e:"
 
@@ -85,8 +86,8 @@ const char *ipfix_elements = DEFAULT_IPFIX_ELEMENTS;
 /* Template Manager */
 struct ipfix_template_mgr *template_mgr = NULL;
 
-/* main loop indicator */
-volatile int done = 0;
+/* terminating indicator */
+volatile int terminating = 0;
 
 /* reconfiguration indicator */
 volatile int reconf = 0;
@@ -105,9 +106,9 @@ int ring_buffer_size = 8192;
  */
 void version ()
 {
-	printf ("%s: IPFIX Collector capture daemon.\n", PACKAGE);
-	printf ("Version: %s, Copyright (C) 2011 CESNET z.s.p.o.\n", VERSION);
-	printf ("See http://www.liberouter.org/technologies/ipfixcol/ for more information.\n\n");
+	printf ("%s: IPFIX Collector capture daemon\n", PACKAGE);
+	printf ("Version: %s, Copyright (C) 2015 CESNET z.s.p.o.\n", VERSION);
+	printf ("Check out http://www.liberouter.org/technologies/ipfixcol/ for more information.\n\n");
 }
 
 /**
@@ -116,16 +117,16 @@ void version ()
 void help ()
 {
 	printf ("Usage: %s [-c file] [-i file] [-dhVs] [-v level]\n", PACKAGE);
-	printf ("  -c file   Path to configuration file (%s by default)\n", DEFAULT_CONFIG_FILE);
-	printf ("  -i file   Path to internal configuration file (%s by default)\n", INTERNAL_CONFIG_FILE);
-	printf ("  -e file   Path to ipfix-elements.xml file (%s by default)\n", DEFAULT_IPFIX_ELEMENTS);
+	printf ("  -c file   Path to configuration file (default: %s)\n", DEFAULT_CONFIG_FILE);
+	printf ("  -i file   Path to internal configuration file (default: %s)\n", INTERNAL_CONFIG_FILE);
+	printf ("  -e file   Path to IPFIX IE specification file (default: %s)\n", DEFAULT_IPFIX_ELEMENTS);
 	printf ("  -d        Daemonize\n");
 	printf ("  -h        Print this help\n");
-	printf ("  -v level  Print verbose messages up to specified level\n");
+	printf ("  -v level  Increase logging verbosity (level: 0-3)\n");
 	printf ("  -V        Print version information\n");
-	printf ("  -s        Skip invalid sequence number error (especially when receiving Netflow v9 data format)\n");
-	printf ("  -r        Ring buffer size\n");
-	printf ("  -S num    Print proccessing statistics every \"num\" seconds\n");
+	printf ("  -s        Skip invalid sequence number error (especially useful for NetFlow v9 PDUs)\n");
+	printf ("  -r        Ring buffer size (default: 8192)\n");
+	printf ("  -S num    Print statistics every \"num\" seconds\n");
 	printf ("\n");
 }
 
@@ -139,12 +140,12 @@ void term_signal_handler(int sig)
 	}
 	
 	/* Terminating signal */
-	if (done) {
-		MSG_COMMON(ICMSG_ERROR, "Another termination signal (%i) detected - quiting without cleanup.", sig);
+	if (terminating) {
+		MSG_COMMON(ICMSG_ERROR, "Another termination signal (%i) detected - quiting without cleanup", sig);
 		exit (EXIT_FAILURE);
 	} else {
-		MSG_COMMON(ICMSG_ERROR, "Signal: %i detected, will exit as soon as possible", sig);
-		done = 1;
+		MSG_COMMON(ICMSG_ERROR, "Signal: %i detected; exiting as soon as possible...", sig);
+		terminating = 1;
 	}
 }
 
@@ -173,6 +174,9 @@ int main (int argc, char* argv[])
 		case 'i':
 			internal_file = optarg;
 			break;
+		case 'e':
+			ipfix_elements = optarg;
+			break;
 		case 'd':
 			daemonize = true;
 			break;
@@ -182,7 +186,13 @@ int main (int argc, char* argv[])
 			exit (EXIT_SUCCESS);
 			break;
 		case 'v':
-			verbose = atoi (optarg);
+			verbose = strtoi (optarg, 10);
+			if (verbose == INT_MAX) {
+				MSG_ERROR(msg_module, "No valid verbosity level provided (%s)", optarg);
+				help ();
+				exit (EXIT_FAILURE);
+			}
+
 			break;
 		case 'V':
 			version ();
@@ -192,16 +202,25 @@ int main (int argc, char* argv[])
 			skip_seq_err = 1;
 			break;
 		case 'r':
-			ring_buffer_size = atoi (optarg);
+			ring_buffer_size = strtoi (optarg, 10);
+			if (ring_buffer_size == INT_MAX) {
+				MSG_ERROR(msg_module, "No valid ring buffer size provided (%s)", optarg);
+				help ();
+				exit (EXIT_FAILURE);
+			}
+
 			break;
 		case 'S':
-			stat_interval = atoi(optarg);
-			break;
-		case 'e':
-			ipfix_elements = optarg;
+			stat_interval = strtoi (optarg, 10);
+			if (stat_interval == INT_MAX) {
+				MSG_ERROR(msg_module, "No valid statistics interval provided (%s)", optarg);
+				help ();
+				exit (EXIT_FAILURE);
+			}
+
 			break;
 		default:
-			MSG_ERROR(msg_module, "Unknown parameter %c", c);
+			MSG_ERROR(msg_module, "Unknown parameter (%c)", c);
 			help ();
 			exit (EXIT_FAILURE);
 			break;
@@ -225,27 +244,24 @@ int main (int argc, char* argv[])
 	 */LIBXML_TEST_VERSION
 	xmlIndentTreeOutput = 1;
 
-	/*
-	 * open and prepare XML configuration file
-	 */
 	/* check config file */
 	if (config_file == NULL) {
 		/* and use default if not specified */
 		config_file = DEFAULT_CONFIG_FILE;
-		MSG_NOTICE(msg_module, "Using default configuration file %s.", config_file);
+		MSG_NOTICE(msg_module, "Using default configuration file: %s", config_file);
 	}
 
 	/* check internal config file */
 	if (internal_file == NULL) {
 		/* and use default if not specified */
 		internal_file = INTERNAL_CONFIG_FILE;
-		MSG_NOTICE(msg_module, "Using default internal configuration file %s.", internal_file);
+		MSG_NOTICE(msg_module, "Using default internal configuration file: %s", internal_file);
 	}
 	  
 	/* Initialize configurator */
 	config = config_init(internal_file, config_file);
 	if (!config) {
-		MSG_ERROR(msg_module, "Configurator initialization failed!");
+		MSG_ERROR(msg_module, "Configurator initialization failed");
 		goto cleanup_err;
 	}
 	
@@ -253,7 +269,7 @@ int main (int argc, char* argv[])
 	collectors = get_collectors(config->act_doc);
 	if (collectors == NULL) {
 		/* no collectingProcess configured */
-		MSG_ERROR(msg_module, "No collectingProcess configured - nothing to do.");
+		MSG_ERROR(msg_module, "No collector process configured");
 		goto cleanup_err;
 	}
 	
@@ -269,13 +285,13 @@ int main (int argc, char* argv[])
 				proc_count++;
 				continue;
 			} else if (pid < 0) { /* error occured, fork failed */
-				MSG_ERROR(msg_module, "Forking collector process failed (%s), skipping collector %d.", strerror(errno), i);
+				MSG_ERROR(msg_module, "Forking collector process failed (%s); skipping collector '%d'", strerror(errno), i);
 				continue;
 			}
 			/* else child - just continue to handle plugins */
 			config->proc_id = i;
 			
-			MSG_NOTICE(msg_module, "[%d] New collector process started.", config->proc_id);
+			MSG_NOTICE(msg_module, "[%d] New collector process started", config->proc_id);
 		}
 		
 		/* DEBUG - remove this */
@@ -287,10 +303,19 @@ int main (int argc, char* argv[])
 		break;
 	}
 
+	/* daemonize */
+	if (daemonize) {
+		closelog();
+		MSG_SYSLOG_INIT(PACKAGE);
+		/* and send all following messages to the syslog */
+		if (daemon (1, 0)) {
+			MSG_ERROR(msg_module, "%s", strerror(errno));
+		}
+	}
+	
 	/*
 	 * create Template Manager
 	 */
-
 	template_mgr = tm_create();
 	if (template_mgr == NULL) {
 		MSG_ERROR(msg_module, "[%d] Unable to create Template Manager", config->proc_id);
@@ -309,33 +334,23 @@ int main (int argc, char* argv[])
 	
 	/* Parse plugins configuration */
 	if (config_reconf(config) != 0) {
-		MSG_ERROR(msg_module, "[%d] Unable to parse plugins configuration", config->proc_id);
+		MSG_ERROR(msg_module, "[%d] Unable to parse plugin configuration", config->proc_id);
 		goto cleanup_err;
-	}
-	
-	/* daemonize */
-	if (daemonize) {
-		closelog();
-		MSG_SYSLOG_INIT(PACKAGE);
-		/* and send all following messages to the syslog */
-		if (daemon (1, 0)) {
-			MSG_ERROR(msg_module, "%s", strerror(errno));
-		}
 	}
 	
 	/* configure output subsystem */
 	retval = output_manager_start();
 	if (retval != 0) {
-		MSG_ERROR(msg_module, "[%d] Initiating Storage Manager failed.", config->proc_id);
+		MSG_ERROR(msg_module, "[%d] Storage Manager initialization failed", config->proc_id);
 		goto cleanup;
 	}
 	
 	/* main loop */
-	while (!done) {
+	while (!terminating) {
 		/* get data to process */
 		if ((get_retval = config->input.get (config->input.config, &input_info, &packet, &source_status)) < 0) {
-			if ((!reconf && !done) || get_retval != INPUT_INTR) { /* if interrupted and closing, it's ok */
-				MSG_WARNING(msg_module, "[%d] Getting IPFIX data failed!", config->proc_id);
+			if ((!reconf && !terminating) || get_retval != INPUT_INTR) { /* if interrupted and closing, it's ok */
+				MSG_WARNING(msg_module, "[%d] Could not get IPFIX data", config->proc_id);
 			}
 			
 			if (reconf) {
@@ -357,7 +372,7 @@ int main (int argc, char* argv[])
 			}
 			/* if input plugin is file reader, end collector */
 			if (input_info->type == SOURCE_TYPE_IPFIX_FILE) {
-				done = 1;
+				terminating = 1;
 			}
 		}
 		/* distribute data to the particular Data Manager for further processing */
@@ -400,9 +415,9 @@ cleanup:
 	if (pid > 0) {
 		for (i=0; i<proc_count; i++) {
 			pid = wait(NULL);
-			MSG_NOTICE(msg_module, "[%d] Collector child process %d terminated", config->proc_id, pid);
+			MSG_NOTICE(msg_module, "[%d] Collector child process '%d' terminated", config->proc_id, pid);
 		}
-		MSG_NOTICE(msg_module, "[%d] Closing collector.", config->proc_id);
+		MSG_NOTICE(msg_module, "[%d] Closing collector", config->proc_id);
 	}
 
 	/* destroy template manager */
@@ -411,7 +426,7 @@ cleanup:
 	}
 
 	xmlCleanupThreads();
-	xmlCleanupParser ();
+	xmlCleanupParser();
 
 	return (retval);
 }