+++ plugins/intermediate/profiler/profile_tree/Profile.cpp
@@ -1,9 +1,9 @@
 /**
- * \file utils.h
+ * \file Profile.cpp
  * \author Michal Kozubik <kozubik@cesnet.cz>
- * \brief Functions unrelated to IPFIX data parsing
+ * \brief intermediate plugin for profiling data
  *
- * Copyright (C) 2014 CESNET, z.s.p.o.
+ * Copyright (C) 2015 CESNET, z.s.p.o.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -37,15 +37,99 @@
  *
  */
 
-#ifndef UTILS_H
-#define	UTILS_H
+#include <algorithm>
 
-#include "api.h"
+#include "Profile.h"
+#include "Channel.h"
 
-API char **utils_files_from_path(char *path);
-API char  *utils_dir_from_path(char *path);
-API char  *strncpy_safe (char *destination, const char *source, size_t num);
-API int    strtoi (const char* str, int base);
+/* Numer of profiles (ID for new profiles) */
+profile_id_t Profile::profiles_cnt = 1;
 
-#endif	/* UTILS_H */
+/**
+ * Constructor
+ */
+Profile::Profile(std::string name)
+: m_id{profiles_cnt++}, m_name{name}
+{
+}
 
+/**
+ * Destructor
+ */
+Profile::~Profile()
+{
+	/* Remove channels */
+	for (auto ch: m_channels) {
+		delete ch;
+	}
+	
+	/* Remove children */
+	for (auto p: m_children) {
+		delete p;
+	}
+}
+
+/**
+ * Add new channel
+ */
+void Profile::addChannel(Channel* channel)
+{
+	m_channels.push_back(channel);
+}
+
+/**
+ * Add child profile
+ */
+void Profile::addProfile(Profile* child)
+{
+	m_children.push_back(child);
+}
+
+/**
+ * Remove child profile
+ */
+void Profile::removeProfile(profile_id_t id)
+{
+	/* Find profile */
+	profilesVec::iterator it = std::find_if(m_children.begin(), m_children.end(),
+										   [id](Profile *p) { return p->getId() == id;});
+
+	/* Remove it */
+	if (it != m_children.end()) {
+		m_children.erase(it);
+	}
+}
+
+/**
+ * Remove channel
+ */
+void Profile::removeChannel(channel_id_t id)
+{
+	/* Find channel */
+	channelsVec::iterator it = std::find_if(m_channels.begin(), m_channels.end(),
+											[id](Channel *ch) { return ch->getId() == id;});
+
+	/* Remove it */
+	if (it == m_channels.end()) {
+		return;
+	}
+
+	/* Unsubscribe channel from it's sources */
+	Channel *ch = *it;
+	for (auto src: ch->getSources()) {
+		src->removeListener(ch);
+	}
+
+	delete ch;
+}
+
+
+/**
+ * Match profile
+ */
+void Profile::match(ipfix_message* msg, metadata* mdata, std::vector<couple_id_t>& profiles)
+{	
+	for (auto channel: m_channels) {
+		channel->match(msg, mdata, profiles);
+	}
+}