+++ plugins/intermediate/profiler/profiler.h
@@ -43,15 +43,13 @@
 extern "C" {
 #include <ipfixcol.h>
 #include "filter.h"
-#include <libxml/parser.h>
-#include <libxml/tree.h>
-#include <libxml/xpathInternals.h>
-#include <libxml2/libxml/parser.h>
-#include <libxml2/libxml/xmlstring.h>
 }
 
-#include "Rule.h"
+/* ID types can by changed here */
+using profile_id_t = uint16_t;
+using channel_id_t = uint16_t;
+using couple_id_t  = uint32_t;
+
 #include <stdexcept>
-#include <vector>
 
 #endif	/* PROFILER_H */