+++ tools/fbitdump/src/Configuration.cpp
@@ -144,11 +144,26 @@ int Configuration::init(int argc, char *argv[])
 			this->resolver = new Resolver(optarg);
 			break;
 		case 'N': /* print plain numbers */
-			if (optarg == std::string("")) {
-				throw std::invalid_argument("-N requires a level specification");
+			if (optarg == NULL || optarg == std::string("")) {
+				/* If the value after '-N' (separated by whitespace) is an integer,
+				 * the user has probably used a whitespace mistakenly.
+				 */
+				if (optind < argc && Utils::strtoi(argv[optind], 10) < INT_MAX) {
+					this->plainLevel = Utils::strtoi(argv[optind], 10);
+
+					// Skip to next argument; make sure integer is not parsed as query filter
+					++optind;
+				} else {
+					this->plainLevel = 1;
+				}
+			} else {
+				this->plainLevel = Utils::strtoi(argv[optind], 10);
+
+				if (this->plainLevel == INT_MAX) {
+					throw std::invalid_argument("-N requires an integer level specification");
+				}
 			}
 
-			this->plainLevel = atoi(optarg);
 			break;
 		case 's': {
 			/* Similar to -A option*/
@@ -163,10 +178,10 @@ int Configuration::init(int argc, char *argv[])
 					throw std::invalid_argument("Argument for option -s is too long");
 				}
 				/* column name is before '/' */
-				Utils::strncpy_safe(parseArg, optarg, pos);
+				Utils::strncpy_safe(parseArg, optarg, pos+1);
 				parseArg[pos] = '\0';
 			} else { /* use whole optarg as column name */
-				Utils::strncpy_safe(parseArg, optarg, arg.size());
+				Utils::strncpy_safe(parseArg, optarg, arg.size()+1);
 			}
 
 			parseAggregateArg(parseArg);
@@ -272,11 +287,15 @@ int Configuration::init(int argc, char *argv[])
 			break;
 		case 'i': /* create indexes */
 			this->createIndexes = true;
-			indexes = optarg;
+			if (optarg != NULL) {
+				indexes = optarg;
+			}
 			break;
 		case 'd': /* delete indexes */
 			this->deleteIndexes = true;
-			indexes = optarg;
+			if (optarg != NULL) {
+				indexes = optarg;
+			}
 			break;
 		case 'C': /* Configuration file */
 			if (optarg == std::string("")) {
@@ -337,7 +356,7 @@ int Configuration::init(int argc, char *argv[])
 		processMOption(tables, optionM.c_str(), optionr);
 	}
 
-	/* allways process option -m, we need to know whether aggregate or not */
+	/* always process option -m, we need to know whether aggregate or not */
 	if (this->optm) {
 		this->processmOption(optionm);
 	}
@@ -349,6 +368,7 @@ int Configuration::init(int argc, char *argv[])
 	Utils::printStatus( "Loading modules");
 
 	this->loadModules();
+
 	/* read filter */
 	if (optind < argc) {
 		this->filter = argv[optind];
@@ -363,8 +383,7 @@ int Configuration::init(int argc, char *argv[])
 		t.seekg(0, std::ios::beg);
 
 		this->filter.assign((std::istreambuf_iterator<char>(t)),
-        std::istreambuf_iterator<char>());
-
+		std::istreambuf_iterator<char>());
 	} else {
 		/* set default filter */
 		this->filter = "1=1";
@@ -787,11 +806,11 @@ void Configuration::help() const
 	/* lines with // at the beginning should be implemented sooner */
 	std::cout
 	<< "usage "<< PACKAGE <<" [options] [\"filter\"]" << std::endl
-	<< "-h              Print help (shown here)" << std::endl
+	<< "-h              Show this help" << std::endl
 	<< "-v <level>      Set verbosity level" << std::endl
 	<< "-V              Print version and exit" << std::endl
 	<< "-a              Aggregate flow data" << std::endl
-	<< "-A [<expr>]     Aggregation fields, separated by ','. For a list of fields, please check fbitdump(1)" << std::endl
+	<< "-A[<expr>]     Aggregation fields, separated by ','. Please check fbitdump(1) for a list of supported fields" << std::endl
 //	<< "                or subnet aggregation: srcip4/24, srcip6/64." << std::endl
 	//<< "-b              Aggregate flow records as bidirectional flows." << std::endl
 	//<< "-B              Aggregate flow records as bidirectional flows - Guess direction." << std::endl
@@ -800,14 +819,14 @@ void Configuration::help() const
 	<< "-n <number>     Define number of top N. -c option takes precedence over -n" << std::endl
 	<< "-c <number>     Limit number of records to display" << std::endl
 	<< "-D <dns>        Use nameserver <dns> for host lookup. Does not support IPv6 addresses" << std::endl
-	<< "-N <level>      Set level of plain numbers printing. For detailed informations, please check fbitdump(1)" << std::endl
+	<< "-N[<level>]     Set plain number printing level. Please check fbitdump(1) for detailed information" << std::endl
 	<< "-s <column>[/<order>]     Generate statistics for <column> any valid record element" << std::endl
 	<< "                and ordered by <order>. Order can be any summarizable column, just as for -m option" << std::endl
-	<< "-q              Quiet: Do not print the header and bottom stat lines" << std::endl
-	<< "-e				Extended bottom stats. Print summary of statistics columns" << std::endl
+	<< "-q              Quiet: do not print statistics" << std::endl
+	<< "-e              Extended statistics: also prints summary of statistics columns" << std::endl
 	//<< "-H Add xstat histogram data to flow file.(default 'no')" << std::endl
-	<< "-i [column1[,column2,...]]	Build indexes for given columns (or all) for specified data" << std::endl
-	<< "-d [column1[,column2,...]]	Delete indexes for given columns (or all) for specified data" << std::endl
+	<< "-i[column1[,column2,...]]	Build indexes for given columns (or all) for specified data" << std::endl
+	<< "-d[column1[,column2,...]]	Delete indexes for given columns (or all) for specified data" << std::endl
 	//<< "-j <file>       Compress/Uncompress file." << std::endl
 	//<< "-z              Compress flows in output file. Used in combination with -w." << std::endl
 	//<< "-l <expr>       Set limit on packets for line and packed output format." << std::endl
@@ -821,9 +840,9 @@ void Configuration::help() const
 	<< "                expr can be dir, which loads the dir from all directories specified in -M," << std::endl
 	<< "				or dir1:dir2, which reads data from subdirectories 'dir1' to 'dir2', in directories from -M" << std::endl
 	<< "-m [column]     Print flow data date sorted. Takes optional parameter '%column' to sort by" << std::endl
-	<< "-R <expr>       Read input from directory (and subdirectories recursively). Can be repeated" << std::endl
-	<< "                /any/dir  Read all data from directory 'dir'." << std::endl
-	<< "                /dir/dir1:dir2: Read all data from directories 'dir1' to 'dir2'" << std::endl
+	<< "-R <expr>       Recursively read input from directory and subdirectories; can be repeated" << std::endl
+	<< "                /any/dir        Reads all data from directory 'dir'" << std::endl
+	<< "                /dir/dir1:dir2  Reads all data from directory 'dir1' to 'dir2'" << std::endl
 	<< "-o <mode>       Use <mode> to print out flow records:" << std::endl
 //	<< "                 raw      Raw record dump." << std::endl
 	<< "                 line     Standard output line format." << std::endl
@@ -832,20 +851,19 @@ void Configuration::help() const
 	<< "                 extra    More than you want to know..." << std::endl
 	<< "                 csv      ',' separated, machine parseable output format" << std::endl
 	<< "                 pipe     '|' separated legacy machine parseable output format" << std::endl
-	<< "                        modes line, long, extended and extra may be extended by '4' '6' to display" << std::endl
-	<< "                        only IPv4 or IPv6 addresses. e.g.long4, extended6" << std::endl
+	<< "                        modes line, long, extended and extra may be extended by '4' or '6' to display" << std::endl
+	<< "                        only IPv4 or IPv6 addresses. Examples: long4, extended6" << std::endl
 //	<< "-v <file>       verify flow data file. Print version and blocks." << std::endl
 	//<< "-x <file>       verify extension records in flow data file." << std::endl
 	//<< "-X              Dump Filtertable and exit (debug option)." << std::endl
-	<< "-Z              Check filters syntax and exit" << std::endl
-	<< "-t <time>       Time window for filtering packets" << std::endl
-	<< "                yyyy/MM/dd.hh:mm:ss[-yyyy/MM/dd.hh:mm:ss]" << std::endl
+	<< "-Z              Check filter syntax and exit" << std::endl
+	<< "-t <time>       Time window for filtering packets: yyyy/MM/dd.hh:mm:ss[-yyyy/MM/dd.hh:mm:ss]" << std::endl
 	<< "-C <path>       Path to configuration file. Default is " << CONFIG_XML << std::endl
 	<< "-T              Print information about templates in directories specified by -R" << std::endl
 	<< "-S              Print available semantics" << std::endl
 	<< "-O              Print available output formats" << std::endl
 	<< "-l              Print plugin list" << std::endl
-	<< "-P <filter>     Post-aggregate filter (only when aggregating, containing columns in aggregated table only)" << std::endl
+	<< "-P <filter>     Post-aggregation filter (only supported with -A, containing columns in aggregated table only)" << std::endl
 	;
 }
 
@@ -979,44 +997,44 @@ void Configuration::processMOption(stringVector &tables, const char *optarg, std
 
 void Configuration::processROption(stringVector &tables, const char *optarg)
 {
-    std::string arg = optarg;
+	std::string arg = optarg;
 
-    /* Find separator */
-    size_t pos = arg.find(':');
-    if (pos == arg.npos) {
-        /* Specified only one directory */
+	/* Find separator */
+	size_t pos = arg.find(':');
+	if (pos == arg.npos) {
+		/* Specified only one directory */
 	Utils::sanitizePath(arg);
-        tables.push_back(arg);
-        return;
-    }
+		tables.push_back(arg);
+		return;
+	}
 
-    /* Get left and right path */
-    std::string left = arg.substr(0, pos);
-    std::string right = arg.substr(pos + 1);
+	/* Get left and right path */
+	std::string left = arg.substr(0, pos);
+	std::string right = arg.substr(pos + 1);
 
-    /* Right path is ready and can be sanitized */
-    Utils::sanitizePath(right);
+	/* Right path is ready and can be sanitized */
+	Utils::sanitizePath(right);
 
-    /* Get root directory (== left - depth of right) */
-    uint16_t right_depth = std::count(right.begin(), right.end(), '/');
+	/* Get root directory (== left - depth of right) */
+	uint16_t right_depth = std::count(right.begin(), right.end(), '/');
 
-    std::string root = left;
-    while (right_depth--) {
-	root = root.substr(0, root.find_last_of('/'));
-    }
+	std::string root = left;
+	while (right_depth--) {
+	   root = root.substr(0, root.find_last_of('/'));
+	}
 
-    /* Get first firectory */
-    if (root == left) {
-	root = "./";
-    } else {
-	left = left.substr(root.length() + 1);
-    }
+	/* Get first firectory */
+	if (root == left) {
+	   root = "./";
+	} else {
+	   left = left.substr(root.length() + 1);
+	}
 
-    Utils::sanitizePath(root);
-    Utils::sanitizePath(left);
+	Utils::sanitizePath(root);
+	Utils::sanitizePath(left);
 
-    /* Load dirs */
-    Utils::loadDirsTree(root, left, right, tables);
+	/* Load dirs */
+	Utils::loadDirsTree(root, left, right, tables);
 }
 
 void Configuration::parseAggregateArg(char *arg) throw (std::invalid_argument)
@@ -1030,7 +1048,7 @@ void Configuration::parseAggregateArg(char *arg) throw (std::invalid_argument)
 	if (arg == NULL) return;
 
 	if (!Utils::splitString(arg, this->aggregateColumnsAliases)) {
-		throw std::invalid_argument(std::string("Ivalid input string ") + arg);
+		throw std::invalid_argument(std::string("Invalid input string '") + arg + std::string("'"));
 	}
 }
 
@@ -1064,7 +1082,7 @@ void Configuration::loadOutputFormat()
 	pugi::xpath_node format = this->getXMLConfiguration().select_single_node(("/configuration/output/format[formatName='"+this->format+"']").c_str());
 	/* check what we found */
 	if (format == NULL) {
-		throw std::invalid_argument(std::string("Format '") + this->format + "' is not defined");
+		throw std::invalid_argument(std::string("Format '") + this->format + "' not defined");
 	}
 
 	/* set format string */