+++ base/changelog.md
@@ -13,3 +13,4 @@
 *  New storage plugin: fastbit_compression
 *  Added reconfiguration at runtime
 *  New intermediate plugin: geoip
+*  Added "terminating" flag accessible from plugins